genai-api-integration/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appopener==1.7
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.13
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==46.0.4
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
deepface==0.0.95
distro==1.9.0
duckduckgo_search==8.1.1
executing==2.2.0
fastapi==0.133.0
fastavro==1.12.1
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
flet==0.80.5
flet-cli==0.80.5
flet-desktop==0.80.5
flet-web==0.80.5
fonttools==4.61.0
fsspec==2026.2.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
groq==1.1.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.8.1
jaxlib==0.8.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.24.2
openai==2.17.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openwakeword==0.6.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow==12.0.0
platformdirs==4.3.7
primp==1.0.0
prompt_toolkit==3.0.50
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
pypylon==26.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==310
PyYAML==6.0.3
pyzmq==26.3.0
qrcode==8.2
regex==2026.1.15
repath==0.9.0
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
SpeechRecognition==3.14.5
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
text-unidecode==1.3
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.24.1
types-requests==2.32.4.20260324
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.41.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==2.0.1
